requirements.txt
Django==3.2; python_version >= "3.6"
django-q==1.3.9
djangorestframework==3.13.1
factory-boy==3.2.1